requirement.txt
ultralytics
opencv-python
Pillow
torch
torchvision
transformers
accelerate
bitsandbytes
streamlit
plotly
pandas
numpy
scikit-learn
scipy
matplotlib